symphony/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "symphony-service"
version = "0.1.0"
description = "Symphony orchestration service implementation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "Jinja2>=3.1",
  "requests>=2.31",
]

[project.scripts]
symphony = "symphony_service.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["symphony_service*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
